changed old to current, new to proposed
