+++ elements/stop-note/src/stop-note-proposed.js
@@ -4,13 +4,14 @@ import { SimpleIconsetStore } from "@lrnwebcomponents/simple-icon/lib/simple-ico
 import { I18NMixin } from "@lrnwebcomponents/i18n-manager/lib/I18NMixin.js";
 import "@lrnwebcomponents/simple-icon/simple-icon.js";
 import { SimpleColors } from "@lrnwebcomponents/simple-colors/simple-colors.js";
+import { DDD } from "@lrnwebcomponents/d-d-d/d-d-d.js";
 
 // register the iconset
 SimpleIconsetStore.registerIconset(
   "stopnoteicons",
   `${
     new URL("./stop-note.js", import.meta.url).href
-  }/../lib/svgs/stopnoteicons/`
+  }/../lib/svgs/stopnoteicons-new/`
 );
 /**
  * `stop-note`
@@ -18,7 +19,7 @@ SimpleIconsetStore.registerIconset(
  * @demo demo/index.html
  * @element stop-note
  */
-class StopNote extends I18NMixin(remoteLinkBehavior(SimpleColors)) {
+class StopNote extends I18NMixin(remoteLinkBehavior(DDD)) {
   /**
    * LitElement constructable styles enhancement
    */
@@ -28,9 +29,11 @@ class StopNote extends I18NMixin(remoteLinkBehavior(SimpleColors)) {
         :host {
           display: block;
           width: auto;
-          --background-color: #f7f7f7;
-          --accent-color: #d32f2f;
-          margin-bottom: 20px;
+          --background-color: var(--ddd-theme-polaris-errorLight);
+          --accent-color: var(--ddd-theme-polaris-error);
+          margin: var(--ddd-spacing-5) 0;
+          font-family: var(--ddd-font-primary);
+          font-weight: var(--ddd-font-primary-bold);
         }
 
         simple-icon {
@@ -39,19 +42,23 @@ class StopNote extends I18NMixin(remoteLinkBehavior(SimpleColors)) {
         }
 
         :host([icon="stopnoteicons:stop-icon"]) {
-          --accent-color: #d8261c;
+          --accent-color: var(--ddd-theme-polaris-original87Pink);
+          --background-color: var(--ddd-theme-polaris-errorLight);
         }
 
         :host([icon="stopnoteicons:warning-icon"]) {
-          --accent-color: #ffeb3b;
+          --accent-color: var(--ddd-theme-polaris-keystoneYellow);
+          --background-color: var(--ddd-theme-polaris-warningLight);
         }
 
         :host([icon="stopnoteicons:confirm-icon"]) {
-          --accent-color: #81c784;
+          --accent-color: var(--ddd-theme-polaris-success);
+          --background-color: var(--ddd-theme-polaris-successLight);
         }
 
         :host([icon="stopnoteicons:book-icon"]) {
-          --accent-color: #21a3db;
+          --accent-color: var(--ddd-theme-polaris-info);
+          --background-color: var(--ddd-theme-polaris-infoLight);
         }
 
         .container {
@@ -59,30 +66,42 @@ class StopNote extends I18NMixin(remoteLinkBehavior(SimpleColors)) {
           width: auto;
         }
 
+        :host([url='']):not([title='']) .message_wrap { /* trying to add more padding to top if there's no link */
+          padding: var(--ddd-spacing-3) var(--ddd-spacing-6);
+        }
+
         .message_wrap {
-          border-right: 7px solid var(--accent-color);
-          padding: 10px 25px;
+          border-right: 4px solid var(--accent-color);
+          padding: var(--ddd-spacing-1) var(--ddd-spacing-6);
           flex: 1 1 auto;
           background-color: var(--background-color);
         }
 
-        .main_message {
-          font-size: 32px;
-          margin-top: 10px;
+        :host([title=""]) .secondary_message {
+          display: flex;
+          height: 90%;
+          align-items: center;
+          margin-top: auto;
+          font-size: var(--ddd-font-size-s);
+          width: 100%;
+          font-weight: var(--ddd-font-primary-bold);
         }
 
         .secondary_message {
-          margin-top: 5px;
-          font-size: 19.2px;
+          font-size: var(--ddd-font-size-3xs);
           width: 100%;
+          font-weight: var(--ddd-font-primary-regular);
         }
 
+
         .link a {
-          margin-top: 5px;
-          font-size: 19.2px;
-          float: left;
-          clear: left;
+          font-size: var(--ddd-font-size-3xs);
+        }
+        a:-webkit-any-link {
           text-decoration: none;
+        }
+        a:hover {
+          text-decoration: underline;
         }
 
         .svg {
@@ -91,9 +110,13 @@ class StopNote extends I18NMixin(remoteLinkBehavior(SimpleColors)) {
         }
 
         .svg_wrap {
-          background-color: var(--accent-color);
-          padding: 5px;
+          background-color: var(--ddd-theme-polaris-limestoneMaxLight);
+          padding: var(--ddd-spacing-1);
           width: auto;
+        }
+
+        .stop-icon {
+          fill: var(--accent-color);
         }
       `,
     ];
@@ -107,14 +130,14 @@ class StopNote extends I18NMixin(remoteLinkBehavior(SimpleColors)) {
           </div>
         </div>
         <div class="message_wrap">
-          <div class="main_message" id="title">${this.title}</div>
-          <div class="secondary_message">
+          <h3 class="main_message mt-1 mb-0 pb-0" id="title">${this.title}</h3>
+          <div class="secondary_message my-2">
             <slot></slot>
             <slot name="message"></slot>
           </div>
           ${this.url
             ? html`
-                <div class="link">
+                <div class="link mb-1">
                   <a href="${this.url}" id="link">
                     ${this.t.moreInformation} &gt;
                   </a>
